requirements.txt
requests
beautifulsoup4
spotipy
selenium
webdriver-manager
python-dotenv
schedule